move convert_units helper into forest.util
